toolchain update for issue #9

+++ config.mk
@@ -1,26 +1,25 @@
+### config.mk
 ### ------------------------------------------------------------ ###
+### Gatemate Makefile for oss-cad-suite toolchain (2026-02-11)   ###
+###                                                              ###
 ### Central Makefile with shared settings across step01...19 The ###
 ### Steps only differ by using additional code modules that are  ###
 ### added into each step folders Makefile.                       ###
 ### ------------------------------------------------------------ ###
 
-## toolchain
-# disable CC-provided Yosys, switch to "OSS CAD Suite" version
-# see https://github.com/fm4dd/gatemate-riscv/issues/8
+## toolchain location
 YOSYS = /home/fm/oss-cad-suite/bin/yosys
-# disable CC-provided openFPGALoader, switch to "OSS CAD Suite" version
-# see https://github.com/fm4dd/gatemate-riscv/issues/5
+PR    = /home/fm/oss-cad-suite/bin/nextpnr-himbaechel
+PACK  = /home/fm/oss-cad-suite/bin/gmpack
 OFL   = /home/fm/oss-cad-suite/bin/openFPGALoader
-PR    = /home/fm/cc-toolchain-linux/bin/p_r/p_r
+VLTR  = /home/fm/oss-cad-suite/bin/verilator
+GTKW  = /home/fm/oss-cad-suite/bin/gtkwave
+IVL   = /home/fm/oss-cad-suite/bin/iverilog
+VVP   = /home/fm/oss-cad-suite/bin/vvp
+IVLFLAGS = -Winfloop -g2012 -gspecify -Ttyp -DSIMULATION
 
-GTKW = gtkwave
-IVL  = iverilog
-VVP  = vvp
-IVLFLAGS = -Winfloop -g2012 -gspecify -Ttyp
-
-## simulation libraries
-CELLS_SYNTH = /home/fm/cc-toolchain-linux/bin/yosys/share/gatemate/cells_sim.v
-CELLS_IMPL = /home/fm/cc-toolchain-linux/bin/p_r/cpelib.v
+## simulation libraries (oss-cad-suite: cpelib.v is now cells_sim.v)
+CELLS_SYNTH = /home/fm/oss-cad-suite/share/yosys/gatemate/cells_sim.v
 
 ## top module, same for each step
 PROJ = SOC
@@ -31,42 +30,86 @@ PIN_DEF = ../gatemate-e1.ccf
 
 all: impl
 
+## --------------------------------------------------------------------------
+## In synthesis we create both output formats: Verilog and JSON.
+## The old Verilog output is needed because iVerlog does not understand JSON.
+## --------------------------------------------------------------------------
 synth_vlog: $(PROJ).v
-	$(YOSYS) -p 'read -sv $(PROJ).v $(ADD_SRC); synth_gatemate -top $(PROJ) -vlog $(PROJ)_synth.v'
+	@test -d log || mkdir log
+	@test -d net || mkdir net
+	$(YOSYS) -ql log/synth.log -p 'read -sv $(PROJ).v $(ADD_SRC); synth_gatemate -top $(PROJ) -luttree -nomx8 -vlog net/$(PROJ)_synth.v; write_json net/$(PROJ)_synth.json'
 
+## --------------------------------------------------------------------------
+## In place_&_route we tell nextpnr to export a Verilog netlist of the design
+## with '--write net/$(PROJ)_impl.v'. This is needed for iVerilog simulations.
+## --------------------------------------------------------------------------
 impl: synth_vlog
 	test -e $(PIN_DEF) || exit
-	$(PR) -i $(PROJ)_synth.v -o $(PROJ) -ccf $(PIN_DEF) +uCIO > $(PROJ)_pr.log
+	$(PR) --device=CCGM1A1 --json net/$(PROJ)_synth.json --write net/$(PROJ)_impl.v -o out=net/$(PROJ)_impl.txt -o ccf=$(PIN_DEF) --router router2 > log/$@.log
+	$(PACK) --input net/$(PROJ)_impl.txt --bit $(PROJ).bit
 
+## --------------------------------------------------------------------------
 ## iVerilog simulation
+## --------------------------------------------------------------------------
 test:
 	@echo 'Running testbench simulation'
 	test ! -e $(PROJ).tb || rm $(PROJ).tb
 	test ! -e $(PROJ).vcd || rm $(PROJ).vcd
-	/usr/bin/iverilog -DBENCH -o $(PROJ).tb -s $(PROJ)_tb $(PROJ)_tb.v $(PROJ).v $(ADD_SRC)
-	/usr/bin/vvp $(PROJ).tb
+	$(IVL) -DBENCH -o $(PROJ).tb -s $(PROJ)_tb $(PROJ)_tb.v $(PROJ).v $(ADD_SRC)
+	$(VVP) $(PROJ).tb
 
+## --------------------------------------------------------------------------
 ## Verilator simulation
+## --------------------------------------------------------------------------
 vtest:
+ifeq ($(wildcard $(PROJ).cpp),)
+	@echo "Error $(PROJ).cpp not found!"
+else
 	@echo 'Running verilator testbench simulation'
 	test ! -d ./obj_dir || rm -rf ./obj_dir
-	test -e $(PROJ).cpp || echo 'Error $(PROJ).cpp not found!'
-	#/usr/bin/verilator --CFLAGS '-I..' -DBENCH -Wno-fatal --top-module $(PROJ) --cc --exe $(PROJ).cpp $(PROJ).v $(ADD_SRC)
-	/usr/bin/verilator -DBENCH -Wno-fatal --top-module $(PROJ) --cc --exe $(PROJ).cpp $(PROJ).v $(ADD_SRC) $(PROJ)_tb.v
-	(cd obj_dir; make -f VSOC.mk)
+	$(VLTR) -DBENCH -Wno-fatal --top-module $(PROJ) --cc --exe $(PROJ).cpp $(PROJ).v $(ADD_SRC) $(PROJ)_tb.v
+	(cd obj_dir; $(MAKE) -f V$(PROJ).mk)
 	test -e obj_dir/V$(PROJ) && obj_dir/V$(PROJ) || echo 'Make failed, no obj_dir/V$(PROJ) found!'
+endif
 
-prog: $(PROJ)_00.cfg
+prog: $(PROJ).bit
 	@echo 'Programming E1 SPI Config:'
-	$(OFL) -b gatemate_evb_spi $<
+	$(OFL) $(OFLFLAGS) -b gatemate_evb_spi $<
 
-flash: $(PROJ)_00.cfg
+flash: $(PROJ).bit
 	@echo 'Programming E1 SPI Flash:'
-	$(OFL) -b gatemate_evb_spi -f --verify $<
+	$(OFL) $(OFLFLAGS) -b gatemate_evb_spi -f --verify $<
 
+
+## -------------------------------------------------------------
+## GTKWave Waveform Generation
+## -------------------------------------------------------------
+
+# Rule to generate VCD from the RTL simulation binary
+$(PROJ).vcd: $(PROJ)_sim.vvp
+	@echo "Generating fresh VCD data for $(PROJ)..."
+	$(VVP) -N $< -lx2
+
+# Primary wave target: check dependencies and launch GTKWave
+wave: $(PROJ).vcd
+	$(GTKW) $< config.gtkw
+
+## -------------------------------------------------------------
+## make clean deletes all compiled files and the bitstream
+## -------------------------------------------------------------
 clean:
-	rm -f $(PROJ)_synth.v $(PROJ)_pr.log $(PROJ)_00.* *.id *.idh *.tb *.prn *.ref* lut*.txt *.idh *.net *.pos *.cdf *.pathes abc.history
-	rm -rf obj_dir
+	$(RM) log/*.log
+	$(RM) net/*_synth.json
+	$(RM) net/*_synth.v
+	$(RM) net/*_impl.v
+	$(RM) net/*_impl.txt
+	$(RM) *.tb
+	$(RM) *.vcd
+	$(RM) *.vvp
+	$(RM) *.bit
+	test ! -d log || rmdir log
+	test ! -d net || rmdir net
+	test ! -d obj_dir || rm -rf obj_dir
 
 .SECONDARY:
 .PHONY: all prog clean
